# openapi.yml
openapi: 3.1.0
info:
  title: Система бронирования мест в кафе
  version: 0.0.4
servers:
  - url: http://localhost:8000/
    description: Локальный сервер
...
paths:
...
  /cafes:
    get:
      tags:
        - Кафе
      summary: Получение списка кафе
      description: Получение списка кафе. Для администраторов и менеджеров - все кафе (с возможностью выбора), для пользователей - только активные.
      operationId: get_all_cafes_cafes_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: show_all
          in: query
          required: false
          schema:
            type: boolean
            title: Показывать все кафе?
            description: Показывать все кафе или нет. По умолчанию показывает все кафе
            default: false
          description: Показывать все кафе или нет. По умолчанию показывает все кафе
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CafeInfo'
                title: Response Get All Cafes Cafes Get
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    post:
      tags:
        - Кафе
      summary: Создание нового кафе
      description: Создает новое кафе. Только для администраторов и менеджеров.
      operationId: create_cafe_cafes_post
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CafeCreate'
              description: Данные кафе
      responses:
        '201':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CafeInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
  /cafes/{cafe_id}:
    get:
      tags:
        - Кафе
      summary: Получение информации о кафе по его ID
      description: Получение информации о кафе по его ID. Для администраторов и менеджеров - все кафе, для пользователей - только активные.
      operationId: get_cafe_by_id_cafes__cafe_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CafeInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
    patch:
      tags:
        - Кафе
      summary: Обновление информации о кафе по его ID
      description: Обновление информации о кафе по его ID. Только для администраторов и менеджеров.
      operationId: path_cafe_by_id_cafes__cafe_id__patch
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: cafe_id
          in: path
          required: true
          schema:
            type: integer
            description: ID кафе
            title: Cafe Id
          description: ID кафе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CafeUpdate'
              description: Обновленные данные кафе
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CafeInfo'
        '400':
          description: Ошибка в параметрах запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '401':
          description: Неавторизированный пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '404':
          description: Данные не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
        '422':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
...
components:
  schemas:
...
    CafeCreate:
      properties:
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        phone:
          type: string
          title: Phone
        description:
          type: string
          title: Description
        photo_id:
          type: string
          format: uuid
          title: Photo Id
        managers_id:
          items:
            type: integer
          type: array
          title: Managers Id
      type: object
      required:
        - name
        - address
        - phone
        - photo_id
        - managers_id
      title: CafeCreate
    CafeInfo:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        address:
          type: string
          title: Address
        phone:
          type: string
          title: Phone
        description:
          type: string
          title: Description
        photo_id:
          type: string
          format: uuid
          title: Photo Id
        managers:
          items:
            $ref: '#/components/schemas/UserShortInfo'
          type: array
          title: Managers
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      title: CafeInfo
...
    CafeUpdate:
      properties:
        name:
          anyOf:
            - type: string
            - type: 'null'
          title: Name
        address:
          anyOf:
            - type: string
            - type: 'null'
          title: Address
        phone:
          anyOf:
            - type: string
            - type: 'null'
          title: Phone
        description:
          anyOf:
            - type: string
            - type: 'null'
          title: Description
        photo_id:
          anyOf:
            - type: string
              format: uuid
            - type: 'null'
          title: Photo Id
        managers_id:
          anyOf:
            - items:
                type: integer
              type: array
            - type: 'null'
          title: Managers Id
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      title: CafeUpdate
    CustomError:
      properties:
        code:
          type: integer
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
        - code
        - message
      title: CustomError
...
    UserShortInfo:
      properties:
        id:
          type: integer
          title: Id
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        phone:
          type: string
          title: Phone
        tg_id:
          type: string
          title: Tg Id
      type: object
      title: UserShortInfo
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/login